pkg/util: move Trie to k8s.io/apiserver
